ci: add deploy workflows
Also includes a nix-based dev env for build tooling, which will install
requirements like `pnpm` and `firebase-tools`.

Opts in to creating a fully static export of the site. Requires that we
disable image optimization, which only works with a server on the
backend.

Uses the firebase-generated github action workflows to deploy to
static-site hosting.