chore: pull in latest penumbra guide
There were a few recent changes made to the guide in the protocol repo,
as of [0], and this commit pulls those changes in. At the very least,
I noticed that changes from the following PRs were missing:

* https://github.com/penumbra-zone/penumbra/pull/4732
* https://github.com/penumbra-zone/penumbra/pull/4726

Tried to make sure to preserve unique content updates implemented
during the mdbook -> nextra conversion.

[0] https://github.com/penumbra-zone/penumbra/tree/a2ffd8a79f1b397e81df7b7e31b256d545cdf004/docs/guide