ci: permit manual deploys from main
We hadn't yet merged the monorepo PR that deactivates the old deploy [0]
logic, so the two fought and the old mdbook site was deployed for a
while. Let's add a config line so we can deploy directly from the web
UI, and not have to push triggering commits, e.g. [1].

[0] https://github.com/penumbra-zone/penumbra/pull/4751
[1] https://github.com/penumbra-zone/guide/commit/b329c697fab4a104cf2be4dbcce7aba398a9b51c