feat: remove hardcoded versions in chain-upgrade docs
Back when the guide was managed in the Penumbra protocol repo [0],
the version strings in the docs were automatically bumped in line with
the source code changes. Since breaking the guide out into a separate
repo, we no longer have the convenience of those automatic updates, so
the guide versions will necessarily lag behind. Let's remove the
hardcoded version strings from crucial reference pages like the
chain-upgrade docs, leaning on node operators to choose the correct
version for a given upgrade.

[0] https://github.com/penumbra-zone/penumbra