docs: add testnet service endpoints
Adds a new dev docs page that provides details on accessing the
currently running testnet. We don't yet have a solution for a faucet, so
the docs will need to be updated to be useful to external developers,
but I'm submitting this PR anyway because this information is necessary,
but not sufficient.

Refs https://github.com/penumbra-zone/guide/issues/35