lint: fix broken links throughout
Adds `markdown-link-check` to CI, and updates all URLs reported as
failing. There are a few exceptions, added to the ignore list by regex.
For the most part, image URLs are substituted during linting so that the
relpath to the file can be found.

In several locations, I updated the nextra-specific `/foo` URLs to be a
more Markdown-compliant `./foo.md`. Not sure whether this will have an
adverse affect on site performance, but I suggest it's a worthwhile
change to prevent broken links from plaguing the docs.

ci: add lint workflow