ci: discrete script for image path checking
Adds a vibe-coded CI script to check that we haven't broken img links.
Previously, we performed a substitution on img src attributes in the
markdown-link-check config; that only worked because all the images
conveniently had the same relpath to the images on disk. Since the
refactor, that's no longer the case. So let's get smart about
inspecting.

Made sure to introduce some typos into img srcs, and the new script
caught it, so I'm calling it good enough.