use recent prax urls throughout
We were previously linking directly to the Chrome Web Store when guiding
users to Prax; there's now a slick homepage for the extension, so let's
link to that everywhere, instead.

Also updated a few of the frontends to refer to recent software names,
but doesn't include actual content for those stubs yet.

Follow-up to excellent review comments in https://github.com/penumbra-zone/guide/pull/79