refactor: reorganize sidebar (#79)
* refactor: reorganize sidebar

Preserves the page content throughout, but heavily modifies the
navigational sidebar, focusing on readability at a glance.

* fix: update broken links

* fix: update redirects for new routes

* ci: discrete script for image path checking

Adds a vibe-coded CI script to check that we haven't broken img links.
Previously, we performed a substitution on img src attributes in the
markdown-link-check config; that only worked because all the images
conveniently had the same relpath to the images on disk. Since the
refactor, that's no longer the case. So let's get smart about
inspecting.

Made sure to introduce some typos into img srcs, and the new script
caught it, so I'm calling it good enough.

* fix: clean up broken links post-refactor

H/t to @TalDerei for comprehensive interaction testing. Caught a few
more broken links that CI didn't catch; fixed!

* refactor: rename using-penumbra -> usage

Makes for terser URLs.